change: 手机网站支付快捷方式由 wap 改为 h5
