+++ src/Shortcut/Unipay/H5Shortcut.php
@@ -2,37 +2,27 @@
 
 declare(strict_types=1);
 
-namespace Yansongda\Pay\Tests\Shortcut\Unipay;
+namespace Yansongda\Pay\Shortcut\Unipay;
 
+use Yansongda\Pay\Contract\ShortcutInterface;
 use Yansongda\Pay\Plugin\ParserPlugin;
 use Yansongda\Pay\Plugin\Unipay\HtmlResponsePlugin;
 use Yansongda\Pay\Plugin\Unipay\LaunchPlugin;
 use Yansongda\Pay\Plugin\Unipay\OnlineGateway\WapPayPlugin;
 use Yansongda\Pay\Plugin\Unipay\PreparePlugin;
 use Yansongda\Pay\Plugin\Unipay\RadarSignPlugin;
-use Yansongda\Pay\Shortcut\Unipay\WapShortcut;
-use Yansongda\Pay\Tests\TestCase;
 
-class WapShortcutTest extends TestCase
+class H5Shortcut implements ShortcutInterface
 {
-    protected WapShortcut $plugin;
-
-    protected function setUp(): void
+    public function getPlugins(array $params): array
     {
-        parent::setUp();
-
-        $this->plugin = new WapShortcut();
-    }
-
-    public function test()
-    {
-        self::assertEquals([
+        return [
             PreparePlugin::class,
             WapPayPlugin::class,
             HtmlResponsePlugin::class,
             RadarSignPlugin::class,
             LaunchPlugin::class,
             ParserPlugin::class,
-        ], $this->plugin->getPlugins([]));
+        ];
     }
 }
