+++ tests/Shortcut/Wechat/H5ShortcutTest.php
@@ -2,9 +2,8 @@
 
 declare(strict_types=1);
 
-namespace Yansongda\Pay\Shortcut\Wechat;
+namespace Yansongda\Pay\Tests\Shortcut\Wechat;
 
-use Yansongda\Pay\Contract\ShortcutInterface;
 use Yansongda\Pay\Plugin\ParserPlugin;
 use Yansongda\Pay\Plugin\Wechat\AddPayloadBodyPlugin;
 use Yansongda\Pay\Plugin\Wechat\AddPayloadSignaturePlugin;
@@ -13,12 +12,23 @@ use Yansongda\Pay\Plugin\Wechat\Pay\H5\PayPlugin;
 use Yansongda\Pay\Plugin\Wechat\ResponsePlugin;
 use Yansongda\Pay\Plugin\Wechat\StartPlugin;
 use Yansongda\Pay\Plugin\Wechat\VerifySignaturePlugin;
+use Yansongda\Pay\Shortcut\Wechat\H5Shortcut;
+use Yansongda\Pay\Tests\TestCase;
 
-class WapShortcut implements ShortcutInterface
+class H5ShortcutTest extends TestCase
 {
-    public function getPlugins(array $params): array
+    protected H5Shortcut $plugin;
+
+    protected function setUp(): void
     {
-        return [
+        parent::setUp();
+
+        $this->plugin = new H5Shortcut();
+    }
+
+    public function testDefault()
+    {
+        self::assertEquals([
             StartPlugin::class,
             PayPlugin::class,
             AddPayloadBodyPlugin::class,
@@ -27,6 +37,6 @@ class WapShortcut implements ShortcutInterface
             ResponsePlugin::class,
             VerifySignaturePlugin::class,
             ParserPlugin::class,
-        ];
+        ], $this->plugin->getPlugins([]));
     }
 }
